Some grammar and speech changes
Added some grammar corrections to the first paragraphs, along with some "personal" touches. Hope it pleases the main dev! :)

Also, if it would not bother too much i suggest we could change our speech to be a bit more informal and relaxed, instead of taking almost everything straight out of the official docs. Just a suggestion!